Abstract away integration failure test commonalities
